backend_sistema/requirements.txt
# Core
flask==3.1.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.1.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Browser automation
playwright==1.49.1

# AI Providers
openai==1.82.0
anthropic==0.52.0

# Export
openpyxl==3.1.5
